chore(workflow): tauri-bundler path considering workspace
`tauri-bundler` is not in the main cargo workspace so the target directory is not in the top level. Update the publishPath to compensate.

Also removing a couple extraneous things that aren't needed either. We had an extra '/' and we don't need to bother showing the directory within the `cargo package` command.