feat(bundler): Sign update archive automatically if pubkey is set
Check in tauri.conf.json if pubkey is set.

Actually this is what an updater configuration may look:

```
"updater": {
        "active": true,
        "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IEY1OTgxQzc0MjVGNjM0Q0IKUldUTE5QWWxkQnlZOWFBK21kekU4OGgzdStleEtkeStHaFR5NjEyRHovRnlUdzAwWGJxWEU2aGYK",
        "endpoints": [
          "http://badurl.www.tld/1",
          "https://gist.githubusercontent.com/lemarier/8e4703d077ebd6470810927ed2205470/raw/329b7ad9f32d439083af40e7f2090ca072d7a1cf/gistfile1.txt?target={{target}}&version={{current_version}}"
        ]
      }
```

TAURI_PRIVATE_KEY is read from environment variable.
It can be a Path or a String

TAURI_KEY_PASSWORD is also an environment variable who can be set (when we bundle) to set the private key password for sining.

If this is a wrong key pasword an error is thrown.

Signed-off-by: David L <david@lemarier.ca>