feat(updater): Validate signature before extraction if pubkey is provided
When we call `updater.install` a new field has been added. It allow to send the pubkey.

If the pubkey is sent, the updater will expect a `signature` field on the update server.

Example:
https://gist.githubusercontent.com/lemarier/72a2a488f1c87601d11ec44d6a7aff05/raw/f86018772318629b3f15dbb3d15679e7651e36f6/with_sign.json

BEFORE extracting we verify the archive and make sure the signature match. If it don’t the update is stopped.

Signed-off-by: David L <david@lemarier.ca>