feat(updater): Add new walk_to_dest when we extract
Now when we extract we keep previous files in the dest we sinmply overwrite with the files in the update archive.

There is a backup made if something failed the dest is restored from the untouched backup and return a failed error inside the updater.  You can prompt the user to download manually the new version by example.

Signed-off-by: David L <david@lemarier.ca>