feat(bundler): Add updater bundling tests
It’s a complete test who make sure the whole process works fine.

We load the signature validator from the updater api as it use a light version of minisign and we need to make sure it’s compatible.

The test is creating the archive, sign it and make sure the signature is valid.

Maybe we can add more test for archive extraction test.

Signed-off-by: David L <david@lemarier.ca>