completely move and build/run examples.
This change is probably overkill, it's a stopgap while i determine the
minimum necessary changes to examples to make them useful to the workspace.
I can't tell if they are actually working, and if my changes so far broke it.

In this commit tests are also not working, only `cargo clippy` is. be warned.