move examples from `tauri/examples/` to `examples/`
This is to allow typechecking examples while developing while still
allowing the examples to be run as part of the workspace and standalone.

The api example is broken for now and will be removed shortly. It can be
added again when the dist files are added to git so that extra build
steps are not needed before running `cargo run --bin api`.