don't remove once listener on new thread
Spawning an async task to remove the once listener caused it to be able to be
called multiple times before being removed. This design choice was previously
made due to deadlock happening when removing the event from inside `fn once`.
That was because the listeners were already locked inside the trigger when
asked to be removed. `fn trigger` now handles removing once handlers