Simplify Args
Remove all generics from `Args` and implement `Params` to it directly.
It also makes signatures of all types and functions based on it
simpler.

This means `Arg` is also the default arg type and it requires `wry`
feature flag. If there's a new runtime in the future, all it needs to do
is add another `Arg` under its feature flag.

Custom command and params should still working, see `params` exmaple to
learn how to define it now. Not much need to change.