doc(api) XDG user dirs are not env vars, fix #3090
This commit points the docs on `path.downloadDir` for XDG_DOWNLOAD_DIR
on Linux, and similar XDG user dirs, at the `xdg-user-dirs` program for
more information, and fixes that they currently imply these are
environment variables; which they a) are unfortunately not by any sort
of standard; b) such env vars are ignored by `dirs-next` which tauri
uses for resolving them.

This does not address the issue of
`${XDG_CONFIG_HOME:-$HOME/.config}/user-dirs.dir` needing to already
exist and contain the appropriate key for this to work (i.e. they do not
default to `$HOME` as the output of `xdg-user-dir` does) - because that
IMO is a shortcoming of `dirs-next` and is better addressed there.