fix: put linux high dpi icons in the correct dir
Note that we are not changing the source paths, "@2x" is still used for those, so this change is not user facing.

On ubuntu, `/usr/share/icons/hicolor` contains dir names that correspond to the new pattern. The previous "@2x" convention comes from Apple, and doesn't seem to apply to Linux.