feat: Use the zbus-backed of notify-rust
While zbus used to have too new MSRV, that is no longer the case for
several months now and hence there is no good reason to avoid zbus
dependency. The MSRV is tracked in the zbus CI even to avoid any
accidental bumps.

OTOH, using zbus eliminates dependency on a platform lib and hence
enhances the portability and ease of building. Issues like these will be a
thing of the past after this change:

https://github.com/tauri-apps/tauri/issues/4373
https://github.com/tauri-apps/tauri/issues/5007
https://github.com/tauri-apps/tauri/issues/5466