fix(runtime-wry): don't hide/show webview when showing/hiding parent
This fixes a regression introduced in #9246 where previously users would wait for `DOMContentLoaded` before showing the window to avoid having a white flash, also introduce a regression for apps that is hidden by default and is shown on global shortcuts or some other action.

Also accoding to the MSDN docs for webview2's `IsVisible` they only recommend to change the webview visibility on when un/minimizing.

closes #9393
ref: https://learn.microsoft.com/en-us/microsoft-edge/webview2/reference/winrt/microsoft_web_webview2_core/corewebview2controller?view=webview2-winrt-1.0.2365.46#isvisible