fix(cli): builtin dev server should also be forwarded for Android
This changes the CLI to persist the "merge config" across mobile command usage (required to pull this config on the xcode and android studio build scripts).
This is required because the builtin dev server makes modifications to the Tauri config (devUrl rewrite) and we need to detect that to forward the dev URL port to Android using ADB.

Also adds a warning on iOS when using `dev --open`: you need the `--host` option to run on physical devices.