feat: Add support for AppLibrary path (#12276)
Previously, there was no access to this specific path used by iOS.

We have added support for this directory, and added documentation
explaining how these different methods are meant to be used.

This also required the creation of a Swift plugin to allow us to access
these iOS APIs.

In this process, we have added a page to the sample views to immediately
see the output of the different path methods.

This required some visual fixes of the examples/api app for the menu button
to be usable on modern Android devices to work with insets.)